@@ -11,18 +11,19 @@
         }
     ],
     "require": {
-        "php": "^5.4|^7.0",
-        "symfony/validator": "^2.3.20|^3.0",
-        "symfony/translation": "^2.3|^3.0",
-        "ronanguilloux/isocodes": "^1.2|^2.0"
+        "php": "^5.6|^7.0",
+        "symfony/validator": "^2.7|^3.0",
+        "symfony/translation": "^2.7|^3.0",
+        "ronanguilloux/isocodes": "^2.0"
     },
     "require-dev": {
-        "symfony/dependency-injection": "^2.3|^3.0",
-        "symfony/http-kernel": "^2.3|^3.0",
-        "symfony/finder": "^2.3|^3.0",
-        "silex/silex": "^1.1",
+        "symfony/dependency-injection": "^2.7|^3.0",
+        "symfony/http-kernel": "^2.7|^3.0",
+        "symfony/finder": "^2.7|^3.0",
+        "silex/silex": "^1.2",
         "matthiasnoback/symfony-dependency-injection-test": "^0.7.4",
         "symfony/phpunit-bridge": "^2.7|^3.0",
+        "sebastian/exporter": "^1.2.0",
         "sllh/php-cs-fixer-styleci-bridge": "^1.0"
     },
     "suggest": {
@@ -32,10 +33,10 @@
         "silex/silex": "For Silex integration as a service provider"
     },
     "conflict": {
-        "symfony/dependency-injection": "<2.3|>=4.0",
-        "symfony/http-kernel": "<2.3|>=4.0",
-        "symfony/finder": "<2.3|>=4.0",
-        "silex/silex": "<1.1.1|>=2.0"
+        "symfony/dependency-injection": "<2.7|>=4.0",
+        "symfony/http-kernel": "<2.7|>=4.0",
+        "symfony/finder": "<2.7|>=4.0",
+        "silex/silex": "<1.2|>=2.0"
     },
     "autoload": {
         "psr-4": { "SLLH\\IsoCodesValidator\\": "src" }
@@ -45,7 +46,7 @@
     },
     "extra": {
         "branch-alias": {
-            "dev-master": "1.x-dev"
+            "dev-master": "2.x-dev"
         }
     }
 }
