@@ -2,38 +2,14 @@
 
 namespace SLLH\IsoCodesValidator\Tests\Constraints;
 
-use SLLH\IsoCodesValidator\Constraints\Cif;
-use SLLH\IsoCodesValidator\Constraints\CifValidator;
-
-class CifValidatorTest extends AbstractConstraintValidatorTest
+/**
+ * @author Sullivan Senechal <soullivaneuh@gmail.com>
+ */
+final class CifValidatorTest extends AbstractGenericConstraintValidatorTest
 {
     /**
-     * @group legacy
+     * {@inheritdoc}
      */
-    protected function setUp()
-    {
-        parent::setUp();
-    }
-
-    protected function createValidator()
-    {
-        return new CifValidator();
-    }
-
-    protected function createConstraint()
-    {
-        return new Cif();
-    }
-
-    /**
-     * @dataProvider getValidValues
-     */
-    public function testValidValues($value)
-    {
-        $this->validator->validate($value, new Cif());
-        $this->assertNoViolation();
-    }
-
     public function getValidValues()
     {
         return [
@@ -49,16 +25,8 @@ class CifValidatorTest extends AbstractConstraintValidatorTest
     }
 
     /**
-     * @dataProvider getInvalidValues
+     * {@inheritdoc}
      */
-    public function testInvalidValues($value)
-    {
-        $this->validator->validate($value, new Cif());
-
-        $this->buildViolation('This value is not a valid CIF.')
-            ->assertRaised();
-    }
-
     public function getInvalidValues()
     {
         return [
@@ -70,7 +38,14 @@ class CifValidatorTest extends AbstractConstraintValidatorTest
             ['N0032484I'],
             ['M0032484I'],
             ['M0032484H'],
-            [' '],
         ];
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    protected function getInvalidMessage()
+    {
+        return 'This value is not a valid CIF.';
     }
 }
