@@ -2,38 +2,14 @@
 
 namespace SLLH\IsoCodesValidator\Tests\Constraints;
 
-use SLLH\IsoCodesValidator\Constraints\Siren;
-use SLLH\IsoCodesValidator\Constraints\SirenValidator;
-
-class SirenValidatorTest extends AbstractConstraintValidatorTest
+/**
+ * @author Sullivan Senechal <soullivaneuh@gmail.com>
+ */
+final class SirenValidatorTest extends AbstractGenericConstraintValidatorTest
 {
     /**
-     * @group legacy
+     * {@inheritdoc}
      */
-    protected function setUp()
-    {
-        parent::setUp();
-    }
-
-    protected function createValidator()
-    {
-        return new SirenValidator();
-    }
-
-    protected function createConstraint()
-    {
-        return new Siren();
-    }
-
-    /**
-     * @dataProvider getValidValues
-     */
-    public function testValidValues($value)
-    {
-        $this->validator->validate($value, new Siren());
-        $this->assertNoViolation();
-    }
-
     public function getValidValues()
     {
         return [
@@ -47,16 +23,8 @@ class SirenValidatorTest extends AbstractConstraintValidatorTest
     }
 
     /**
-     * @dataProvider getInvalidValues
+     * {@inheritdoc}
      */
-    public function testInvalidValues($value)
-    {
-        $this->validator->validate($value, new Siren());
-
-        $this->buildViolation('This value is not a valid SIREN.')
-            ->assertRaised();
-    }
-
     public function getInvalidValues()
     {
         return [
@@ -70,8 +38,15 @@ class SirenValidatorTest extends AbstractConstraintValidatorTest
             ['334932721'],
             ['440288372'],
             ['517439544'],
-            [' '],
             ['azertyuio'],
         ];
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    protected function getInvalidMessage()
+    {
+        return 'This value is not a valid SIREN.';
     }
 }
