@@ -2,38 +2,14 @@
 
 namespace SLLH\IsoCodesValidator\Tests\Constraints;
 
-use SLLH\IsoCodesValidator\Constraints\Siret;
-use SLLH\IsoCodesValidator\Constraints\SiretValidator;
-
-class SiretValidatorTest extends AbstractConstraintValidatorTest
+/**
+ * @author Sullivan Senechal <soullivaneuh@gmail.com>
+ */
+final class SiretValidatorTest extends AbstractGenericConstraintValidatorTest
 {
     /**
-     * @group legacy
+     * {@inheritdoc}
      */
-    protected function setUp()
-    {
-        parent::setUp();
-    }
-
-    protected function createValidator()
-    {
-        return new SiretValidator();
-    }
-
-    protected function createConstraint()
-    {
-        return new Siret();
-    }
-
-    /**
-     * @dataProvider getValidValues
-     */
-    public function testValidValues($value)
-    {
-        $this->validator->validate($value, new Siret());
-        $this->assertNoViolation();
-    }
-
     public function getValidValues()
     {
         return [
@@ -48,16 +24,8 @@ class SiretValidatorTest extends AbstractConstraintValidatorTest
     }
 
     /**
-     * @dataProvider getInvalidValues
+     * {@inheritdoc}
      */
-    public function testInvalidValues($value)
-    {
-        $this->validator->validate($value, new Siret());
-
-        $this->buildViolation('This value is not a valid SIRET.')
-            ->assertRaised();
-    }
-
     public function getInvalidValues()
     {
         return [
@@ -70,9 +38,16 @@ class SiretValidatorTest extends AbstractConstraintValidatorTest
             ['33493272000018'],
             ['44028837100015'],
             ['51743954300012'],
-            [' '],
             ['azertyuiopqsdf'],
 
         ];
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    protected function getInvalidMessage()
+    {
+        return 'This value is not a valid SIRET.';
     }
 }
