@@ -2,38 +2,14 @@
 
 namespace SLLH\IsoCodesValidator\Tests\Constraints;
 
-use SLLH\IsoCodesValidator\Constraints\StructuredCommunication;
-use SLLH\IsoCodesValidator\Constraints\StructuredCommunicationValidator;
-
-class StructuredCommunicationValidatorTest extends AbstractConstraintValidatorTest
+/**
+ * @author Sullivan Senechal <soullivaneuh@gmail.com>
+ */
+final class StructuredCommunicationValidatorTest extends AbstractGenericConstraintValidatorTest
 {
     /**
-     * @group legacy
+     * {@inheritdoc}
      */
-    protected function setUp()
-    {
-        parent::setUp();
-    }
-
-    protected function createValidator()
-    {
-        return new StructuredCommunicationValidator();
-    }
-
-    protected function createConstraint()
-    {
-        return new StructuredCommunication();
-    }
-
-    /**
-     * @dataProvider getValidValues
-     */
-    public function testValidValues($value)
-    {
-        $this->validator->validate($value, new StructuredCommunication());
-        $this->assertNoViolation();
-    }
-
     public function getValidValues()
     {
         return [
@@ -44,16 +20,8 @@ class StructuredCommunicationValidatorTest extends AbstractConstraintValidatorTe
     }
 
     /**
-     * @dataProvider getInvalidValues
+     * {@inheritdoc}
      */
-    public function testInvalidValues($value)
-    {
-        $this->validator->validate($value, new StructuredCommunication());
-
-        $this->buildViolation('This value is not a valid structured communication code.')
-            ->assertRaised();
-    }
-
     public function getInvalidValues()
     {
         return [
@@ -64,7 +32,14 @@ class StructuredCommunicationValidatorTest extends AbstractConstraintValidatorTe
             [10132748107],
             [1013274810067],
             [101374810060],
-            [' '],
         ];
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    protected function getInvalidMessage()
+    {
+        return 'This value is not a valid structured communication code.';
     }
 }
