@@ -2,38 +2,14 @@
 
 namespace SLLH\IsoCodesValidator\Tests\Constraints;
 
-use SLLH\IsoCodesValidator\Constraints\Uknin;
-use SLLH\IsoCodesValidator\Constraints\UkninValidator;
-
-class UkninValidatorTest extends AbstractConstraintValidatorTest
+/**
+ * @author Sullivan Senechal <soullivaneuh@gmail.com>
+ */
+final class UkninValidatorTest extends AbstractGenericConstraintValidatorTest
 {
     /**
-     * @group legacy
+     * {@inheritdoc}
      */
-    protected function setUp()
-    {
-        parent::setUp();
-    }
-
-    protected function createValidator()
-    {
-        return new UkninValidator();
-    }
-
-    protected function createConstraint()
-    {
-        return new Uknin();
-    }
-
-    /**
-     * @dataProvider getValidValues
-     */
-    public function testValidValues($value)
-    {
-        $this->validator->validate($value, new Uknin());
-        $this->assertNoViolation();
-    }
-
     public function getValidValues()
     {
         return [
@@ -44,16 +20,8 @@ class UkninValidatorTest extends AbstractConstraintValidatorTest
     }
 
     /**
-     * @dataProvider getInvalidValues
+     * {@inheritdoc}
      */
-    public function testInvalidValues($value)
-    {
-        $this->validator->validate($value, new Uknin());
-
-        $this->buildViolation('This value is not a valid NINO.')
-            ->assertRaised();
-    }
-
     public function getInvalidValues()
     {
         return [
@@ -63,8 +31,15 @@ class UkninValidatorTest extends AbstractConstraintValidatorTest
             ['AF123456C'],
             ['AB123456F'],
             ['TN011258F'],
-            [' '],
             ['azertyuiop'],
         ];
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    protected function getInvalidMessage()
+    {
+        return 'This value is not a valid NINO.';
     }
 }
