@@ -4,7 +4,7 @@ namespace SLLH\IsoCodesValidator\Tests\Constraints;
 
 use SLLH\IsoCodesValidator\Constraints\ZipCode;
 
-class ZipCodeTest extends \PHPUnit_Framework_TestCase
+final class ZipCodeTest extends \PHPUnit_Framework_TestCase
 {
     /**
      * @dataProvider getValidCountries
@@ -27,34 +27,6 @@ class ZipCodeTest extends \PHPUnit_Framework_TestCase
     }
 
     /**
-     * 1.0 BC.
-     *
-     * @dataProvider getLegacyValidCountries
-     */
-    public function testLegacyValidCountries($country)
-    {
-        $zipCode = new ZipCode(['country' => $country]);
-
-        // 1.0 BC
-        $deprecatedOptionsBridge = [
-            'Canada'        => 'CA',
-            'France'        => 'FR',
-            'Netherlands'   => 'NL',
-        ];
-
-        $this->assertSame($deprecatedOptionsBridge[$country], $zipCode->country);
-    }
-
-    public function getLegacyValidCountries()
-    {
-        return [
-            ['Canada'],
-            ['France'],
-            ['Netherlands'],
-        ];
-    }
-
-    /**
      * @dataProvider getInvalidCountries
      * @expectedException \Symfony\Component\Validator\Exception\ConstraintDefinitionException
      */
@@ -69,17 +41,7 @@ class ZipCodeTest extends \PHPUnit_Framework_TestCase
             ['TL'],
             [42],
             ['Liberlands'],
-            // 1.0 BC Test
-            ['canada'],
-            ['CANADA'],
-            ['caNada'],
-            ['france'],
-            ['FRANCE'],
-            ['frAnce'],
-            ['netherlands'],
-            ['NETHERLANDS'],
-            ['neTherlands'],
-            // END 1.0 BC Test
+            ['France'],
             ['us'],
             ['uS'],
             ['Us'],
