@@ -5,16 +5,11 @@ namespace SLLH\IsoCodesValidator\Tests\Constraints;
 use SLLH\IsoCodesValidator\Constraints\ZipCode;
 use SLLH\IsoCodesValidator\Constraints\ZipCodeValidator;
 
-class ZipCodeValidatorTest extends AbstractConstraintValidatorTest
+final class ZipCodeValidatorTest extends AbstractConstraintValidatorTest
 {
     protected function createValidator()
     {
         return new ZipCodeValidator();
-    }
-
-    protected function createConstraint()
-    {
-        return new ZipCode();
     }
 
     /**
@@ -30,6 +25,9 @@ class ZipCodeValidatorTest extends AbstractConstraintValidatorTest
         $this->assertNoViolation();
     }
 
+    /**
+     * {@inheritdoc}
+     */
     public function getValidValues()
     {
         return [
@@ -59,49 +57,6 @@ class ZipCodeValidatorTest extends AbstractConstraintValidatorTest
     }
 
     /**
-     * 1.0 BC tests.
-     *
-     * @dataProvider getLegacyValidValues
-     */
-    public function testLegacyValidValues($value, $country)
-    {
-        $this->validator->validate($value, new ZipCode([
-            'country' => $country,
-        ]));
-        $this->assertNoViolation();
-        $this->validator->validate($value, new ZipCode());
-        $this->assertNoViolation();
-    }
-
-    public function getLegacyValidValues()
-    {
-        return [
-            ['A0A 1A0', 'Canada'],
-            ['A0A1A0', 'Canada'],
-            ['H0H 0H0', 'Canada'],
-            ['A0A 1A0', 'Canada'],
-            ['06000', 'France'],
-            ['56000', 'France'],
-            ['56420', 'France'],
-            ['20000', 'France'],
-            ['97114', 'France'],
-            ['99999', 'France'],
-            ['99123', 'France'],
-            ['98000', 'France'],
-            ['00100', 'France'],
-            ['01000', 'France'],
-            ['1234AA', 'Netherlands'],
-            ['1234 AA', 'Netherlands'],
-            ['1023 AA', 'Netherlands'],
-            ['99801', 'US'],
-            ['02115', 'US'],
-            ['10001', 'US'],
-            ['20008', 'US'],
-            ['99950', 'US'],
-        ];
-    }
-
-    /**
      * @dataProvider getInvalidValues
      */
     public function testInvalidValues($value, $country)
@@ -110,10 +65,13 @@ class ZipCodeValidatorTest extends AbstractConstraintValidatorTest
             'country' => $country,
         ]));
 
-        $this->buildViolation('This value is not a valid ZIP code.')
+        $this->buildViolation($this->getInvalidMessage())
             ->assertRaised();
     }
 
+    /**
+     * {@inheritdoc}
+     */
     public function getInvalidValues()
     {
         return [
@@ -179,91 +137,14 @@ class ZipCodeValidatorTest extends AbstractConstraintValidatorTest
             ['Lorem Ipsum', 'all'],
             ['LOREM IPSUM', 'all'],
             ['lorem ipsum', 'all'],
-            [' ', 'all'],
         ];
     }
 
     /**
-     * 1.0 BC tests.
-     *
-     * @dataProvider getLegacyInvalidValues
+     * {@inheritdoc}
      */
-    public function testLegacyInvalidValues($value, $country)
+    protected function getInvalidMessage()
     {
-        $this->validator->validate($value, new ZipCode([
-            'country' => $country,
-        ]));
-
-        $this->buildViolation('This value is not a valid ZIP code.')
-            ->assertRaised();
-    }
-
-    public function getLegacyInvalidValues()
-    {
-        return [
-            ['560', 'Canada'],
-            ['5600', 'Canada'],
-            ['560000', 'Canada'],
-            ['A56000', 'Canada'],
-            ['A5600', 'Canada'],
-            ['56000A', 'Canada'],
-            ['A5600A', 'Canada'],
-            ['AAA', 'Canada'],
-            ['AAAA', 'Canada'],
-            ['AAAAA', 'Canada'],
-            ['A 0A1A0', 'Canada'],
-            ['A0 A1A0', 'Canada'],
-            ['A0A1 A0', 'Canada'],
-            ['A0A1A 0', 'Canada'],
-            ['A0A1a0', 'Canada'],
-            ['a0a1a0', 'Canada'],
-            ['2A004', 'France'],
-            ['560', 'France'],
-            ['5600', 'France'],
-            ['560000', 'France'],
-            ['A56000', 'France'],
-            ['A5600', 'France'],
-            ['56000A', 'France'],
-            ['A5600A', 'France'],
-            ['AAA', 'France'],
-            ['AAAA', 'France'],
-            ['AAAAA', 'France'],
-            ['1234', 'Netherlands'],
-            ['1234A', 'Netherlands'],
-            ['AA1234', 'Netherlands'],
-            ['A1234A', 'Netherlands'],
-            ['1A2A3A', 'Netherlands'],
-            ['1234ABC', 'Netherlands'],
-            ['123AB', 'Netherlands'],
-            ['123456', 'Netherlands'],
-            ['AAAA', 'Netherlands'],
-            ['ABCD12', 'Netherlands'],
-            ['1234 ABC', 'Netherlands'],
-            ['12345A', 'Netherlands'],
-            ['1234 5A', 'Netherlands'],
-            ['0123 AA', 'Netherlands'],
-            ['1234aa', 'Netherlands'],
-            ['5600', 'US'],
-            ['560000', 'US'],
-            ['A56000', 'US'],
-            ['A5600', 'US'],
-            ['56000A', 'US'],
-            ['A5600A', 'US'],
-            ['AAA', 'US'],
-            ['AAAA', 'US'],
-            ['AAAAA', 'US'],
-            ['A 0A1A0', 'US'],
-            ['A0 A1A0', 'US'],
-            ['A0A1 A0', 'US'],
-            ['A0A1A 0', 'US'],
-            ['A0A1a0', 'US'],
-            ['a0a1a0', 'US'],
-            ['AAA', 'all'],
-            ['AAAAA', 'all'],
-            ['Lorem Ipsum', 'all'],
-            ['LOREM IPSUM', 'all'],
-            ['lorem ipsum', 'all'],
-            [' ', 'all'],
-        ];
+        return 'This value is not a valid ZIP code.';
     }
 }
